Internationalized most of the main tab, also added paging and the expanded detail view
The expanded detail view is a view that is launched by clicking the bottom right hand corner of the Patient aag view. It will be used to display all information related to a patient, and in the future will be expanded to use with the other entities in the system.

Internationalized most of the main screen.

Also, added the paging of search results.