Reworked security questions.
Security questions are now stored in a seperate table in the
database.  They can be removed or deleted through the
SecurityQuestionDao.  They also have a many to one relationship
with users.

Furthermore, gui layout for the login screen has been removed from
the source coude and instead relegated to the direcgtory:

        src/main/resources/ui/plugins/patientview/login/