Created SecurityOptions class for persistance.
Hashed out a rough framework for accessing and changing security
settings (on the code side of things).  The class is not persistant
yet, but will be soon.  All of the storage was moved out of the
SecurityPanelController and into the singleton SecurityOptions.

Also, Range, formerly an inner class, now has its own file.